Module Tezos_base

module Block_header : sig ... end

module Block_locator : sig ... end

A locator t is a data structure which roughly represents a list of block hashes in the chain. These hashes go from the top of the chain to the bottom. It is sparse in the sense that the distance between two hashes increases exponentially when we move away from the head.

module Data_encoding_wrapper : sig ... end

module Distributed_db_version : sig ... end

Distributed_db protocol version.

module Fitness : sig ... end

module Genesis : sig ... end

module Mempool : sig ... end

Tezos Shell Module - Mempool, a.k.a. the operations safe to be broadcast.

module Network_version : sig ... end

module Operation : sig ... end

module P2p_addr : sig ... end

module P2p_connection : sig ... end

module P2p_identity : sig ... end

module P2p_peer : sig ... end

module P2p_peer_id : sig ... end

module P2p_point : sig ... end

module P2p_rejection : sig ... end

module P2p_stat : sig ... end

Bandwidth usage statistics

module P2p_version : sig ... end

P2p-layer protocol version.

module Point_parser : sig ... end

module Preapply_result : sig ... end

module Protocol : sig ... end

module S : sig ... end

module Test_chain_status : sig ... end

module Time : sig ... end

Time management

module TzPervasives : sig ... end

module User_activated : sig ... end

module Worker_intf : sig ... end

module Worker_types : sig ... end