Module TzPervasives.String

include module type of Stdlib.String

val length : string -> int

val get : string -> int -> char

val set : bytes -> int -> char -> unit

val create : int -> bytes

val make : int -> char -> string

val init : int -> (int -> char) -> string

val copy : string -> string

val sub : string -> int -> int -> string

val fill : bytes -> int -> int -> char -> unit

val blit : string -> int -> bytes -> int -> int -> unit

val concat : string -> string list -> string

val iter : (char -> unit) -> string -> unit

val iteri : (int -> char -> unit) -> string -> unit

val map : (char -> char) -> string -> string

val mapi : (int -> char -> char) -> string -> string

val trim : string -> string

val escaped : string -> string

val index : string -> char -> int

val index_opt : string -> char -> int option

val rindex : string -> char -> int

val rindex_opt : string -> char -> int option

val index_from : string -> int -> char -> int

val index_from_opt : string -> int -> char -> int option

val rindex_from : string -> int -> char -> int

val rindex_from_opt : string -> int -> char -> int option

val contains : string -> char -> bool

val contains_from : string -> int -> char -> bool

val rcontains_from : string -> int -> char -> bool

val uppercase : string -> string

val lowercase : string -> string

val capitalize : string -> string

val uncapitalize : string -> string

val uppercase_ascii : string -> string

val lowercase_ascii : string -> string

val capitalize_ascii : string -> string

val uncapitalize_ascii : string -> string

type t = string

val compare : t -> t -> int

val equal : t -> t -> bool

val split_on_char : char -> string -> string list

val to_seq : t -> char Stdlib.Seq.t

val to_seqi : t -> (int * char) Stdlib.Seq.t

val of_seq : char Stdlib.Seq.t -> t

val unsafe_get : string -> int -> char

val unsafe_set : bytes -> int -> char -> unit

val unsafe_blit : string -> int -> bytes -> int -> int -> unit

val unsafe_fill : bytes -> int -> int -> char -> unit

include module type of Tezos_stdlib.TzString

module Set : Stdlib.Set.S with type elt = string

module Map : Stdlib.Map.S with type key = string

val split_path : string -> string list

Splits a string on slashes, grouping multiple slashes, and ignoring slashes at the beginning and end of string.

val split : char -> ?dup:bool -> ?limit:int -> string -> string list

Splits a string on a delimiter character. If dup is set to true, groups multiple delimiters and strips delimiters at the beginning and end of string. If limit is passed, stops after limit split(s). dups defaults to true and limit defaults to max_int. Examples:

split ~dup:true ',' ",hello,,world,"] returns "hello"; "world"

split ~dup:false ',' ",,hello,,,world,,"] returns ""; "hello"; ""; ""; "world"; ""

val has_prefix : prefix:string -> string -> bool

true if input has prefix

val remove_prefix : prefix:string -> string -> string option

Some (input with prefix removed), if string has prefix, else None

val common_prefix : string -> string -> int

Length of common prefix of input strings

val mem_char : string -> char -> bool

Test whether a string contains a given character

val fold_left : ('a -> char -> 'a) -> 'a -> string -> 'a

Functional iteration over the characters of a string from first to last

val is_hex : string -> bool

Test whether a string is a valid hexadecimal value

val pp_bytes_hex : Stdlib.Format.formatter -> bytes -> unit

Pretty print bytes as hexadecimal string.

module Hashtbl : Tezos_error_monad.TzLwtreslib.Hashtbl.SeededS with type key = t