Module Tezos_base.Worker_intf

Parameters to build a worker group

module type NAME = sig ... end

The name of the group of workers corresponding to an instantiation of Make, as well as the name of each worker in that group.

module type VIEW = sig ... end

module type EVENT = sig ... end

Events that are used for logging and introspection. Events are pretty printed immediately in the log, and stored in the worker's event backlog for introspection.

module type REQUEST = sig ... end

The type of messages that are fed to the worker's event loop.

module type TYPES = sig ... end

The (imperative) state of the event loop.

module type LOGGER = sig ... end