Module Tezos_base__Block_locator

A locator t is a data structure which roughly represents a list of block hashes in the chain. These hashes go from the top of the chain to the bottom. It is sparse in the sense that the distance between two hashes increases exponentially when we move away from the head.

The distance between two hashes of a locator is randomized to prevent attacks. The seed is determined uniquely from the peer_id of the sender and the receiver so that the distance between two hashes can be recomputed locally. This is the purpose of the function to_steps.

The step representation is mostly used by the peer_validator and the bootstrap_pipeline modules.

The last step of a locator may be truncated. It is the case when the last step hits the caboose. Thus, such a non-strict step can be terminated by:

The genesis: it is the case in both Archive and Full modes as the caboose is always located down to the genesis block, but it is also the case in a Rolling mode early state (when caboose = genesis).

Any block: it is the case in Rolling mode when the caboose is higher than the genesis. Indeed, the caboose can be located (almost) anywhere.