Module Tezos_base__Time

Time management

This module supports two distinct notions of time. The first notion of time is the time as handled by the protocol. This is the time that appears in the header of blocks, the time that baking slots are specified on, etc. It only has second-level precision.

The second notion of time is the time as handled by the system. This is the time as returned by the processor clock, the time that network timeouts are specified on, etc. In has sub-second precision.

The distinction between the two notions of time is important for multiple reasons:

Protocol time and system time may evolve independently. E.g., if a protocol update changes the notion of time.

Protocol time and system time have different levels of precision.

Protocol time and system time have different end-of-times. Respectively that's int64 end-of-time (some time in the year 292277026596) and RFC3339 end-of-time (end of the year 9999).

Note that while Protocol time has the int64 range, many of its functions do not work outside of the RFC3339 range, namely:

all xx_notation_xx functions (will be renamed and moved to an RFC submodule later)

rfc_encoding

pp_hum