Module Tezos_base__TzPervasives.Circular_buffer

type t

Type of circular buffers

type data

An abstraction over a chunk of data written in the buffer.

val create : ?⁠maxlength:int -> ?⁠fresh_buf_size:int -> unit -> t

Create a buffer of size maxlength. If the buffer is full, a buffer of size fresh_buf_size is allocated (by default `2` kb).

val write : maxlen:int -> fill_using:(Stdlib.Bytes.t -> int -> int -> int Lwt.t) -> t -> data Lwt.t

write ~maxlen ~fill_using:f buffer calls fill_using buf offset maxlen where buf is a buffer that has room for maxlen data starting from offset.

Assumes that fill_using returns the exact amount of written bytes.

Behaviour is unspecified if fill_using writes more than maxlen data or lies on the number of written bytes.

It returns a data descriptor for the supposedly written chunk.

val read : data -> ?⁠len:int -> t -> into:Stdlib.Bytes.t -> offset:int -> data option

read data ~len ~into:buf buffer ~offset blit len data from the data chunk. If len is not provided, blit all the data. If len is less than the amount of data available, it returns a new handler the the remainder.

Assumes that data has been produce by a write attempt in buffer. Assumes that ~len is lesser than length data.

val length : data -> int

length data returns the amount of avalaible bytes in data