Module type Tezos_base__Worker_intf.NAME

The name of the group of workers corresponding to an instantiation of Make, as well as the name of each worker in that group.

val base : string list

The name/path of the worker group

type t

The abstract name of a single worker

val encoding : t Data_encoding.t

Serializer for the introspection RPCs

val pp : Stdlib.Format.formatter -> t -> unit

Pretty printer for displaying the worker name

val equal : t -> t -> bool