Module type Tezos_base__Worker_intf.REQUEST

The type of messages that are fed to the worker's event loop.

type 'a t

The type of events. It is possible to wait for an event to be processed from outside the worker using push_request_and_wait. In this case, the handler for this event can return a value. The parameter is the type of this value.

include VIEW

type view

val encoding : view Data_encoding.t

val pp : Stdlib.Format.formatter -> view -> unit

val view : 'a t -> view

The projection function from full request to simple views.