Module type Tezos_base__Worker_intf.TYPES

The (imperative) state of the event loop.

type state

The internal state that is passed to the event handlers.

type parameters

The parameters provided when launching a new worker.

type view

A simplified view of the worker's state for introspection.

val view : state -> parameters -> view

The projection function from full state to simple views.

val encoding : view Data_encoding.t

Serializer for the introspection RPCs

val pp : Stdlib.Format.formatter -> view -> unit

Pretty printer for introspection.