Module Tezos_base__Worker_types

type limits = {

backlog_size : int;
Number of event stored in the backlog for each debug level.

backlog_level : Tezos_event_logging.Internal_event.level;
Stores events at least as important as this value.

}

Some memory and time limits.

type worker_status =

| Launching of Tezos_base.Time.System.t

| Running of Tezos_base.Time.System.t

| Closing of Tezos_base.Time.System.t * Tezos_base.Time.System.t

| Closed of Tezos_base.Time.System.t * Tezos_base.Time.System.t * Tezos_error_monad.Error_monad.error list option

The running status of an individual worker.

val worker_status_encoding : Tezos_error_monad.Error_monad.error list Data_encoding.t -> worker_status Data_encoding.t

Worker status serializer for RPCs.

type worker_information = {

instances_number : int;

wstatus : worker_status;

queue_length : int;

}

val worker_information_encoding : Tezos_error_monad.Error_monad.error list Data_encoding.t -> worker_information Data_encoding.t

type request_status = {

pushed : Tezos_base.Time.System.t;

treated : Tezos_base.Time.System.Span.t;

completed : Tezos_base.Time.System.Span.t;

}

The running status of an individual request.

val request_status_encoding : request_status Data_encoding.t

Request status serializer for RPCs.

type ('req, 'evt) full_status = {

status : worker_status;

pending_requests : (Tezos_base.Time.System.t * 'req) list;

backlog : (Tezos_event_logging.Internal_event.level * 'evt list) list;

current_request : (Tezos_base.Time.System.t * Tezos_base.Time.System.t * 'req) option;

}

The full status of an individual worker.

val full_status_encoding : 'req Data_encoding.t -> 'evt Data_encoding.t -> Tezos_error_monad.Error_monad.error list Data_encoding.t -> ('req, 'evt) full_status Data_encoding.t

Full worker status serializer for RPCs.

val pp_status : Stdlib.Format.formatter -> request_status -> unit