Module System.Span

type t = Ptime.Span.t

A representation of spans of time between two timestamps.

val multiply_exn : float -> t -> t

multiply_exn factor t is a time spans that lasts factor time as long as t. It fails if the time span cannot be represented.

val of_seconds_exn : float -> t

of_seconds_exn f is a time span of f seconds. It fails if the time span cannot be represented.

Serialization functions

val rpc_arg : t Tezos_rpc.RPC_arg.t

val pp_hum : Stdlib.Format.formatter -> t -> unit

val encoding : t Data_encoding.t