Module Tezos_base.User_activated

type upgrades = (Stdlib.Int32.t * Tezos_crypto.Protocol_hash.t) list

User activated upgrades: at given level, switch to given protocol.

val upgrades_encoding : upgrades Data_encoding.t

type protocol_overrides = (Tezos_crypto.Protocol_hash.t * Tezos_crypto.Protocol_hash.t) list

User activated protocol overrides.

An override (a, b) denotes that if a is to be activated at the end of the promotion phase, b shall be activated instead.

val protocol_overrides_encoding : protocol_overrides Data_encoding.t